+++ lib/cisco_node_utils/bgp_af.rb
@@ -448,6 +448,79 @@ module Cisco
     end
 
     #
+    # Distance (Getter/Setter/Default)
+    #
+    def distance_set(ebgp, ibgp, local)
+      if ebgp == default_distance_ebgp && ibgp == default_distance_ibgp &&
+         local == default_distance_local
+        set_args_keys(state: 'no', ebgp: distance_ebgp, ibgp: distance_ibgp,
+                      local: distance_local)
+      else
+        set_args_keys(state: '', ebgp: ebgp, ibgp: ibgp, local: local)
+      end
+      config_set('bgp_af', 'distance', @set_args)
+    end
+
+    def distance_ebgp
+      ebgp, _ibgp, _local = distance
+      return default_distance_ebgp if ebgp.nil?
+      ebgp.to_i
+    end
+
+    def distance_ibgp
+      _ebgp, ibgp, _local = distance
+      return default_distance_ibgp if ibgp.nil?
+      ibgp.to_i
+    end
+
+    def distance_local
+      _ebgp, _ibgp, local = distance
+      return default_distance_local if local.nil?
+      local.to_i
+    end
+
+    def distance
+      match = config_get('bgp_af', 'distance', @get_args)
+      match.nil? ? default_distance : match
+    end
+
+    def default_distance_ebgp
+      config_get_default('bgp_af', 'distance_ebgp')
+    end
+
+    def default_distance_ibgp
+      config_get_default('bgp_af', 'distance_ibgp')
+    end
+
+    def default_distance_local
+      config_get_default('bgp_af', 'distance_local')
+    end
+
+    def default_distance
+      ["#{default_distance_ebgp}", "#{default_distance_ibgp}",
+       "#{default_distance_local}"]
+    end
+
+    #
+    # default_metric (Getter/Setter/Default)
+    #
+
+    # default_metric
+    def default_metric
+      config_get('bgp_af', 'default_metric', @get_args)
+    end
+
+    def default_metric=(val)
+      set_args_keys(state: (val) ? '' : 'no',
+                    num:   (val) ? val : default_metric)
+      config_set('bgp_af', 'default_metric', @set_args)
+    end
+
+    def default_default_metric
+      config_get_default('bgp_af', 'default_metric')
+    end
+
+    #
     # maximum_paths (Getter/Setter/Default)
     #
 
@@ -649,6 +722,61 @@ module Cisco
           config_set('bgp_af', prop, @set_args)
         end
       end
+    end
+
+    #
+    # Suppress Inactive (Getter/Setter/Default)
+    #
+    def suppress_inactive
+      state = config_get('bgp_af', 'suppress_inactive', @get_args)
+      state ? true : false
+    end
+
+    def suppress_inactive=(state)
+      state = (state ? '' : 'no')
+      set_args_keys(state: state)
+      config_set('bgp_af', 'suppress_inactive', @set_args)
+    end
+
+    def default_suppress_inactive
+      config_get_default('bgp_af', 'suppress_inactive')
+    end
+
+    #
+    # Table Map (Getter/Setter/Default)
+    #
+
+    def table_map
+      config_get('bgp_af', 'table_map', @get_args)
+    end
+
+    def table_map_filter
+      config_get('bgp_af', 'table_map_filter', @get_args)
+    end
+
+    def table_map_set(map, filter=false)
+      if filter
+        feature = 'table_map_filter'
+      else
+        feature = 'table_map'
+      end
+      if map == default_table_map
+        @set_args[:state] = 'no'
+        @set_args[:map] = table_map
+      else
+        @set_args[:state] = ''
+        @set_args[:map] = map
+      end
+      config_set('bgp_af', feature, @set_args)
+      set_args_keys_default
+    end
+
+    def default_table_map
+      config_get_default('bgp_af', 'table_map')
+    end
+
+    def default_table_map_filter
+      config_get_default('bgp_af', 'table_map_filter')
     end
   end
 end